fix: u64

+++ common/Inc/myints.h
@@ -8,7 +8,7 @@
 typedef uint8_t u8;
 typedef uint16_t u16;
 typedef uint32_t u32;
-typedef uint32_t u64;
+typedef uint64_t u64;
 
 #ifdef USE_IXX
 typedef int8_t i8;
